Remove traces of linting tests
Long time ago linters were run as tests. And ember-exam was splitting
tests into partitions in lint aware style because linting tests are much
faster than regular tests.

Nowadays we don't run linters as tests and this feature is no longer
necessary. Remove it from the code to avoid the confusion of the future
generations.

The initial driver for this change was a failing test which was
complaining on something related to jshint and potentially it was caused
by removal of jshint support from ember-cli.