Drop mocha support
ember-mocha is not compatible with ember@5 and its support is dropped,
so, we can remove mocha support from ember-exam now.

@embroider/macros is not needed anymore as we can use static imports now
everywhere.

qunit and ember-qunit are required peer dependencies now.

Normalize test modules to use the same syntax: tests are defined in the
scope of module, use non-arrow functions in modules and tests definitions.

Normalize `ember-qunit` name everywhere in the docs.